Moved more helper methods out to the utility classes.
