+++ README.md
@@ -143,13 +143,13 @@ TODO
 + *Author*: [Llewellyn van der Merwe](mailto:llewellyn@joomlacomponentbuilder.com)
 + *Name*: [Component Builder](https://github.com/vdm-io/Joomla-Component-Builder)
 + *First Build*: 30th April, 2015
-+ *Last Build*: 10th March, 2022
++ *Last Build*: 12th March, 2022
 + *Version*: 2.12.15
 + *Copyright*: Copyright (C) 2015 Vast Development Method. All rights reserved.
 + *License*: GNU General Public License version 2 or later; see LICENSE.txt
-+ *Line count*: **286315**
++ *Line count*: **286346**
 + *Field count*: **1580**
-+ *File count*: **1848**
++ *File count*: **1850**
 + *Folder count*: **260**
 
 > This **component** was build with a [Joomla](https://extensions.joomla.org/extension/component-builder/) [Automated Component Builder](http://joomlacomponentbuilder.com).
